Cleanup (#25)
* remove cmd folder as we are going to have only one entrypoint

* get rid of pkg directory

* rename test -> integration_tests as unit tests should reside next to the actual files they are testing

* database migration WIP

* reinstate gorm boilerplate in the addinvoice for now to make it compile

* introduce migrations

* add Makefile

* don't use unsigned types for database mappings

* migrations work now

* add build target

* use echo groups

* gorm removed

* add envconfig

* fix comments

+++ db/models/account.go
@@ -2,6 +2,6 @@ package models
 
 // Account : Account Model
 type Account struct {
-	UserID uint
+	UserID int64
 	Type   string
 }
